Improve help commands
Help is now a seperate app, and is a lot more flexible. /help and //help
now displays a list of all commands using ManualListView instead of
through ingame chat. It also improves on how a command is displayed,
such as displaying additional info like command description and
different aliases when applicable.

Further improvements will be to add a description to all commands which
has no description.